refactor: Reorganize the processing pipeline code (#3206)
- move code closer to where it is used.
- move common code from dataset_processing.process.py  to dataset_processing.common.py
- move cxg processing code from dataset_processing.process.py to dataset_processing.process_cxg.py
- move seurat processing code from dataset_processing.process.py to dataset_processing.process_seurat.py
- move logger from dataset_processing.process.py to dataset_processing.logger.py
- remove `dataset_processing.process.process`. It was only used in testing
- split tests from test.unit.backend.corpora.dataset_processing.test_processing to separate files.
- remove `dataset_processing.process.create_artifacts`. It was only used in testing.
- remove check_env. It was only used in testing
- Update tests to work after removing test code.
