+++ backend/corpora/dataset_processing/logger.py
@@ -1,0 +1,4 @@
+import logging
+
+logger = logging.getLogger("dataset_processing")
+logging.basicConfig(level=logging.INFO)
