+++ backend/corpora/dataset_processing/process.py
@@ -133,157 +133,19 @@ state that mimics the Conversion step of the main step function.
 
 """
 
-import logging
 import os
-import subprocess
 import sys
-import typing
-from datetime import datetime
-from os.path import join
 
-import numpy
-import scanpy
-
-from backend.corpora.common.corpora_orm import (
-    ConversionStatus,
-    DatasetArtifactFileType,
-    ProcessingStatus,
-    ValidationStatus,
-    UploadStatus,
-)
-from backend.corpora.common.entities import Dataset, DatasetAsset
-from backend.corpora.common.utils.db_helpers import processing_status_updater
+from backend.corpora.common.entities import Dataset
 from backend.corpora.common.utils.db_session import db_session_manager
-from backend.corpora.common.utils.dl_sources.url import from_url
-from backend.corpora.common.utils.s3_buckets import buckets
-from backend.corpora.dataset_processing.download import download
+from backend.corpora.dataset_processing.common import update_db
 from backend.corpora.dataset_processing.exceptions import (
     ProcessingCancelled,
     ProcessingFailed,
     ValidationFailed,
     ConversionFailed,
 )
-from backend.corpora.dataset_processing.h5ad_data_file import H5ADDataFile
-
-logger = logging.getLogger(__name__)
-logging.basicConfig(level=logging.INFO)
-
-# This is unfortunate, but this information doesn't appear to live anywhere
-# accessible to the uploader
-DEPLOYMENT_STAGE_TO_URL = {
-    "staging": "https://cellxgene.staging.single-cell.czi.technology/e",
-    "prod": "https://cellxgene.cziscience.com/e",
-    "rdev": f"https:/{os.environ.get('REMOTE_DEV_PREFIX')}-explorer.rdev.single-cell.czi.technology/e",
-    "dev": "https://cellxgene.dev.single-cell.czi.technology/e",
-    "test": "http://frontend.corporanet.local:3000",
-}
-
-LABELED_H5AD_FILENAME = "local.h5ad"
-
-
-def check_env():
-    """Verify that the required environment variables are set."""
-
-    missing = []
-    for env_var in [
-        "DROPBOX_URL",
-        "ARTIFACT_BUCKET",
-        "CELLXGENE_BUCKET",
-        "DATASET_ID",
-        "DEPLOYMENT_STAGE",
-    ]:
-        if env_var not in os.environ:
-            missing.append(env_var)
-    if missing:
-        raise EnvironmentError(f"Missing environment variables: {missing}")
-
-
-def create_artifact(
-    file_name: str,
-    artifact_type: DatasetArtifactFileType,
-    bucket_prefix: str,
-    dataset_id: str,
-    artifact_bucket: str,
-    processing_status_type: str,
-):
-    update_db(
-        dataset_id,
-        processing_status={processing_status_type: ConversionStatus.UPLOADING},
-    )
-    logger.info(f"Uploading [{dataset_id}/{file_name}] to S3 bucket: [{artifact_bucket}].")
-    try:
-        s3_uri = DatasetAsset.upload(file_name, bucket_prefix, artifact_bucket)
-        with db_session_manager() as session:
-            logger.info(f"Updating database with  {artifact_type}.")
-            DatasetAsset.create(
-                session,
-                dataset_id=dataset_id,
-                filename=file_name,
-                filetype=artifact_type,
-                user_submitted=True,
-                s3_uri=s3_uri,
-            )
-        update_db(
-            dataset_id,
-            processing_status={processing_status_type: ConversionStatus.UPLOADED},
-        )
-
-    except Exception as e:
-        logger.error(e)
-        e.args = {processing_status_type: ConversionStatus.FAILED}
-        raise e
-
-
-def replace_artifact(
-    file_name: str,
-    bucket_prefix: str,
-    artifact_bucket: str,
-):
-    logger.info(f"Uploading [{bucket_prefix}/{file_name}] to S3 bucket: [{artifact_bucket}].")
-    DatasetAsset.upload(file_name, bucket_prefix, artifact_bucket)
-
-
-def create_artifacts(
-    local_filename: str,
-    dataset_id: str,
-    artifact_bucket: str,
-    can_convert_to_seurat: bool = False,
-):
-    bucket_prefix = get_bucket_prefix(dataset_id)
-    logger.info(f"Creating artifacts for dataset {dataset_id}...")
-    # upload AnnData
-    create_artifact(
-        local_filename,
-        DatasetArtifactFileType.H5AD,
-        bucket_prefix,
-        dataset_id,
-        artifact_bucket,
-        "h5ad_status",
-    )
-
-    if can_convert_to_seurat:
-        # Convert to Seurat and upload
-        seurat_filename = convert_file(
-            make_seurat,
-            local_filename,
-            "Issue creating seurat.",
-            dataset_id,
-            "rds_status",
-        )
-        if seurat_filename:
-            create_artifact(
-                seurat_filename,
-                DatasetArtifactFileType.RDS,
-                bucket_prefix,
-                dataset_id,
-                artifact_bucket,
-                "rds_status",
-            )
-    else:
-        update_db(dataset_id, processing_status=dict(rds_status=ConversionStatus.SKIPPED))
-        logger.info(f"Skipped Seurat conversion for dataset {dataset_id}")
-
-    logger.info(f"Finished creating artifacts for dataset {dataset_id}")
+from backend.corpora.dataset_processing.logger import logger
 
 
 def cancel_dataset(dataset_id):
@@ -292,335 +154,6 @@ def cancel_dataset(dataset_id):
         dataset.asset_deletion()
         dataset.delete()
         logger.info("Upload Canceled.")
-
-
-def update_db(dataset_id, metadata: dict = None, processing_status: dict = None):
-    with db_session_manager() as session:
-        dataset = Dataset.get(session, dataset_id, include_tombstones=True)
-        if dataset.tombstone:
-            raise ProcessingCancelled
-
-        if metadata:
-            logger.debug("Updating metadata.")
-            dataset.update(**metadata)
-
-        if processing_status:
-            logger.debug(f"updating processing_status.{processing_status}")
-            processing_status_updater(session, dataset.processing_status.id, processing_status)
-
-
-def download_from_source_uri(dataset_id: str, source_uri: str, local_path: str) -> str:
-    """Given a source URI, download it to local_path.
-    Handles fixing the url so it downloads directly.
-    """
-
-    file_url = from_url(source_uri)
-    if not file_url:
-        raise ValueError(f"Malformed source URI: {source_uri}")
-
-    # This is a bit ugly and should be done polymorphically instead, but Dropbox support will be dropped soon
-    if file_url.scheme == "https":
-        file_info = file_url.file_info()
-        status = download(dataset_id, file_url.url, local_path, file_info["size"])
-        logger.info(status)
-    elif file_url.scheme == "s3":
-        bucket_name = file_url.netloc
-        key = remove_prefix(file_url.path, "/")
-        wrapped_download_from_s3(
-            dataset_id=dataset_id,
-            bucket_name=bucket_name,
-            object_key=key,
-            local_filename=local_path,
-        )
-    else:
-        raise ValueError(f"Download for URI scheme '{file_url.scheme}' not implemented")
-
-    logger.info("Download complete")
-    return local_path
-
-
-# TODO: after upgrading to Python 3.9, replace this with removeprefix()
-def remove_prefix(string: str, prefix: str) -> str:
-    if string.startswith(prefix):
-        return string[len(prefix) :]
-    else:
-        return string[:]
-
-
-def download_from_s3(bucket_name: str, object_key: str, local_filename: str):
-    logger.info(f"Downloading file {local_filename} from bucket {bucket_name} with object key {object_key}")
-    buckets.portal_client.download_file(bucket_name, object_key, local_filename)
-
-
-def wrapped_download_from_s3(dataset_id: str, bucket_name: str, object_key: str, local_filename: str):
-    """
-    Wraps download_from_s3() to update the dataset's upload status
-    :param dataset_id:
-    :param bucket_name:
-    :param object_key:
-    :param local_filename:
-    :return:
-    """
-    with db_session_manager() as session:
-        processing_status = Dataset.get(session, dataset_id).processing_status
-        processing_status.upload_status = UploadStatus.UPLOADING
-        download_from_s3(
-            bucket_name=bucket_name,
-            object_key=object_key,
-            local_filename=local_filename,
-        )
-        processing_status.upload_status = UploadStatus.UPLOADED
-
-
-def extract_metadata(filename) -> dict:
-    """Pull metadata out of the AnnData file to insert into the dataset table."""
-
-    adata = scanpy.read_h5ad(filename, backed="r")
-
-    # TODO: Concern with respect to previous use of raising error when there is no raw layer.
-    # This new way defaults to adata.X.
-    if adata.raw is not None and adata.raw.X is not None:
-        layer_for_mean_genes_per_cell = adata.raw.X
-    else:
-        layer_for_mean_genes_per_cell = adata.X
-
-    # For mean_genes_per_cell, we only want the columns (genes) that have a feature_biotype of `gene`,
-    # as opposed to `spike-in`
-    filter_gene_vars = numpy.where(adata.var.feature_biotype == "gene")[0]
-
-    # Calling np.count_nonzero on and h5py.Dataset appears to read the entire thing
-    # into memory, so we need to chunk it to be safe.
-    stride = 50000
-    numerator, denominator = 0, 0
-    for bounds in zip(
-        range(0, layer_for_mean_genes_per_cell.shape[0], stride),
-        range(stride, layer_for_mean_genes_per_cell.shape[0] + stride, stride),
-    ):
-        chunk = layer_for_mean_genes_per_cell[bounds[0] : bounds[1], filter_gene_vars]
-        numerator += chunk.nnz if hasattr(chunk, "nnz") else numpy.count_nonzero(chunk)
-        denominator += chunk.shape[0]
-
-    def _get_term_pairs(base_term):
-        base_term_id = base_term + "_ontology_term_id"
-        return [
-            {"label": k[0], "ontology_term_id": k[1]}
-            for k in adata.obs.groupby([base_term, base_term_id]).groups.keys()
-        ]
-
-    def _get_is_primary_data():
-        is_primary_data = adata.obs["is_primary_data"]
-        if all(is_primary_data):
-            return "PRIMARY"
-        elif not any(is_primary_data):
-            return "SECONDARY"
-        else:
-            return "BOTH"
-
-    def _get_x_approximate_distribution():
-        if "X_approximate_distribution" in adata.uns:
-            return adata.uns["X_approximate_distribution"].upper()
-        else:
-            return None
-
-    def _get_batch_condition():
-        if "batch_condition" in adata.uns:
-            return adata.uns["batch_condition"]
-        else:
-            return None
-
-    metadata = {
-        "name": adata.uns["title"],
-        "organism": _get_term_pairs("organism"),
-        "tissue": _get_term_pairs("tissue"),
-        "assay": _get_term_pairs("assay"),
-        "disease": _get_term_pairs("disease"),
-        "sex": _get_term_pairs("sex"),
-        "ethnicity": _get_term_pairs("ethnicity"),
-        "development_stage": _get_term_pairs("development_stage"),
-        "cell_count": adata.shape[0],
-        "mean_genes_per_cell": numerator / denominator,
-        "is_primary_data": _get_is_primary_data(),
-        "cell_type": _get_term_pairs("cell_type"),
-        "x_normalization": adata.uns["X_normalization"],
-        "x_approximate_distribution": _get_x_approximate_distribution(),
-        "schema_version": adata.uns["schema_version"],
-        "batch_condition": _get_batch_condition(),
-    }
-    logger.info(f"Extract metadata: {metadata}")
-    return metadata
-
-
-def make_seurat(local_filename):
-    """Create a Seurat rds file from the AnnData file."""
-
-    try:
-        subprocess.run(
-            [
-                "Rscript",
-                os.path.join(os.path.abspath(os.path.dirname(__file__)), "make_seurat.R"),
-                local_filename,
-            ],
-            capture_output=True,
-            check=True,
-        )
-    except subprocess.CalledProcessError as ex:
-        msg = f"Seurat conversion failed: {ex.output} {ex.stderr}"
-        logger.exception(msg)
-        raise RuntimeError(msg) from ex
-
-    return local_filename.replace(".h5ad", ".rds")
-
-
-def make_cxg(local_filename):
-    """
-    Convert the uploaded H5AD file to the CXG format servicing the cellxgene Explorer.
-    """
-
-    cxg_output_container = local_filename.replace(".h5ad", ".cxg")
-    try:
-        h5ad_data_file = H5ADDataFile(local_filename, var_index_column_name="feature_name")
-        h5ad_data_file.to_cxg(cxg_output_container, sparse_threshold=25.0)
-    except Exception as ex:
-        msg = "CXG conversion failed."
-        logger.exception(msg)
-        raise RuntimeError(msg) from ex
-
-    return cxg_output_container
-
-
-def copy_cxg_files_to_cxg_bucket(cxg_dir, s3_uri):
-    """
-    Copy cxg files to the cellxgene bucket (under the given object key) for access by the explorer
-    """
-    command = ["aws"]
-    if os.getenv("BOTO_ENDPOINT_URL"):
-        command.append(f"--endpoint-url={os.getenv('BOTO_ENDPOINT_URL')}")
-
-    command.extend(
-        [
-            "s3",
-            "cp",
-            cxg_dir,
-            s3_uri,
-            "--recursive",
-            "--acl",
-            "bucket-owner-full-control",
-        ]
-    )
-    subprocess.run(
-        command,
-        check=True,
-    )
-
-
-def convert_file(
-    converter: typing.Callable,
-    local_filename: str,
-    error_message: str,
-    dataset_id: str,
-    processing_status_type: str,
-) -> str:
-    logger.info(f"Converting {local_filename}")
-    start = datetime.now()
-    try:
-        update_db(
-            dataset_id,
-            processing_status={processing_status_type: ConversionStatus.CONVERTING},
-        )
-        file_dir = converter(local_filename)
-        update_db(
-            dataset_id,
-            processing_status={processing_status_type: ConversionStatus.CONVERTED},
-        )
-        logger.info(f"Finished converting {converter} in {datetime.now()- start}")
-    except Exception as e:
-        logger.exception(f"{error_message}: {e}")
-        status = {processing_status_type: ConversionStatus.FAILED}
-        raise ConversionFailed(status)
-    return file_dir
-
-
-def get_bucket_prefix(identifier):
-    remote_dev_prefix = os.environ.get("REMOTE_DEV_PREFIX", "")
-    if remote_dev_prefix:
-        return join(remote_dev_prefix, identifier).strip("/")
-    else:
-        return identifier
-
-
-def process_cxg(local_filename, dataset_id, cellxgene_bucket):
-    cxg_dir = convert_file(make_cxg, local_filename, "Issue creating cxg.", dataset_id, "cxg_status")
-    if cxg_dir:
-        with db_session_manager() as session:
-            asset = DatasetAsset.create(
-                session,
-                dataset_id=dataset_id,
-                filename="explorer_cxg",
-                filetype=DatasetArtifactFileType.CXG,
-                user_submitted=True,
-                s3_uri="",
-            )
-            asset_id = asset.id
-            bucket_prefix = get_bucket_prefix(asset_id)
-            s3_uri = f"s3://{cellxgene_bucket}/{bucket_prefix}.cxg/"
-        update_db(dataset_id, processing_status={"cxg_status": ConversionStatus.UPLOADING})
-        copy_cxg_files_to_cxg_bucket(cxg_dir, s3_uri)
-        metadata = {
-            "explorer_url": join(
-                DEPLOYMENT_STAGE_TO_URL[os.environ["DEPLOYMENT_STAGE"]],
-                dataset_id + ".cxg",
-                "",
-            )
-        }
-        with db_session_manager() as session:
-            logger.info(f"Updating database with cxg artifact for dataset {dataset_id}. s3_uri is {s3_uri}")
-            asset = DatasetAsset.get(session, asset_id)
-            asset.update(s3_uri=s3_uri)
-        update_db(dataset_id, processing_status={"cxg_status": ConversionStatus.UPLOADED})
-
-    else:
-        metadata = None
-    update_db(dataset_id, metadata)
-
-
-def validate_h5ad_file_and_add_labels(dataset_id: str, local_filename: str) -> typing.Tuple[str, bool]:
-    """
-    Validates and labels the specified dataset file and updates the processing status in the database
-    :param dataset_id: ID of the dataset to update
-    :param local_filename: file name of the dataset to validate and label
-    :return: file name of labeled dataset, boolean indicating if seurat conversion is possible
-    """
-    from cellxgene_schema import validate
-
-    update_db(
-        dataset_id,
-        processing_status=dict(validation_status=ValidationStatus.VALIDATING),
-    )
-    output_filename = LABELED_H5AD_FILENAME
-    is_valid, errors, can_convert_to_seurat = validate.validate(local_filename, output_filename)
-
-    if not is_valid:
-        logger.error(f"Validation failed with {len(errors)} errors!")
-        status = dict(
-            validation_status=ValidationStatus.INVALID,
-            validation_message=errors,
-        )
-        raise ValidationFailed(status)
-    else:
-        logger.info("Validation complete")
-        status = dict(
-            h5ad_status=ConversionStatus.CONVERTED,
-            validation_status=ValidationStatus.VALID,
-        )
-        update_db(dataset_id, processing_status=status)
-        return output_filename, can_convert_to_seurat
-
-
-def clean_up_local_file(local_filename):
-    try:
-        os.remove(local_filename)
-    except Exception:
-        pass
 
 
 def log_batch_environment():
@@ -640,30 +173,6 @@ def log_batch_environment():
     for var in batch_environment_variables:
         env_vars[var] = os.getenv(var)
     logger.info(f"Batch Job Info: {env_vars}")
-
-
-def process(dataset_id, dropbox_url, cellxgene_bucket, artifact_bucket):
-    update_db(dataset_id, processing_status=dict(processing_status=ProcessingStatus.PENDING))
-    local_filename = download_from_source_uri(
-        dataset_id=dataset_id,
-        source_uri=dropbox_url,
-        local_path="raw.h5ad",
-    )
-
-    # No file cleanup needed due to docker run-time environment.
-    # To implement proper cleanup, tests/unit/backend/corpora/dataset_processing/test_process.py
-    # will have to be modified since it relies on a shared local file
-
-    file_with_labels, can_convert_to_seurat = validate_h5ad_file_and_add_labels(dataset_id, local_filename)
-
-    # Process metadata
-    metadata = extract_metadata(file_with_labels)
-    update_db(dataset_id, metadata)
-
-    # create artifacts
-    process_cxg(file_with_labels, dataset_id, cellxgene_bucket)
-    create_artifacts(file_with_labels, dataset_id, artifact_bucket, can_convert_to_seurat)
-    update_db(dataset_id, processing_status=dict(processing_status=ProcessingStatus.SUCCESS))
 
 
 def main():
