+++ tests/unit/processing_container/test_process.py
@@ -4,8 +4,9 @@ from unittest.mock import patch
 
 import requests
 
-from backend.corpora.dataset_processing import process
+from backend.corpora.dataset_processing.process import main
 from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
+from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup
 
 
 class TestDatasetProcessing(CorporaTestCaseUsingMockAWS):
@@ -51,23 +52,42 @@ class TestDatasetProcessing(CorporaTestCaseUsingMockAWS):
             if os.path.exists(f):
                 os.remove(f)
 
-    @patch("backend.corpora.dataset_processing.process.download_from_source_uri")
-    def test_main(self, mock_download_from_source_uri):
+    @patch("backend.corpora.dataset_processing.process_download_validate.download_from_source_uri")
+    @patch("backend.corpora.dataset_processing.remaster_cxg.process")  # TODO: provide test data to properly test this.
+    def test_main(self, mock_rematser_cxg_process, mock_download_from_source_uri):
         """
         Tests full pipeline for processing an uploaded H5AD file, including database updates
         generation and upload of all artifacts to S3 (localstack), but excluding the Dropbox download
         functionality.  Dropbox I/O is mocked to prevent dependency on remote services (non-Dockerized).
         """
         mock_download_from_source_uri.return_value = self.h5ad_raw
-
         dataset = self.generate_dataset(self.session, collection_id="test_collection_id")
+        test_environment = {
+            "STEP_NAME": "cxg",
+            "DROPBOX_URL": "https://www.dropbox.com/IGNORED",
+            "ARTIFACT_BUCKET": self.corpora_config.bucket_name,
+            "CELLXGENE_BUCKET": self.corpora_config.bucket_name,
+            "DATASET_ID": dataset.id,
+            "DEPLOYMENT_STAGE": "test",
+            "AWS_BATCH_JOB_ATTEMPT": "1",
+        }
 
-        process.process(
-            dataset.id,
-            "https://www.dropbox.com/IGNORED",
-            self.corpora_config.bucket_name,
-            self.corpora_config.bucket_name,
-        )
+        test_environment["STEP_NAME"] = "download-validate"
+        with EnvironmentSetup(test_environment):
+            main()
+
+        test_environment["STEP_NAME"] = "cxg"
+        with EnvironmentSetup(test_environment):
+            main()
+
+        test_environment["STEP_NAME"] = "seurat"
+        with EnvironmentSetup(test_environment):
+            main()
+
+        test_environment["STEP_NAME"] = "cxg_remaster"
+        with EnvironmentSetup(test_environment):
+            main()
+            mock_rematser_cxg_process.assert_called()
 
         # TODO: add assertions. See https://app.zenhub.com/workspaces/single-cell-5e2a191dad828d52cc78b028/issues/chanzuckerberg/single-cell-data-portal/1449 # noqa: E501
         # 1. H5AD has annotation labels added and uploaded to S3
